requirements.txt
openai==0.27.7
python-dotenv==1.0.0
requests==2.31.0
pytest==7.3.1
requests-mock==1.10.0